+++ src/main/java/sopt/org/homepage/aboutsopt/AboutSoptController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import sopt.org.homepage.aboutsopt.dto.GetAboutSoptResponseDto;
+import sopt.org.homepage.aboutsopt.service.AboutSoptService;
 
 @RestController
 @RequiredArgsConstructor
