+++ src/main/java/sopt/org/homepage/project/ProjectController.java
@@ -9,6 +9,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import sopt.org.homepage.common.dto.PaginateResponseDto;
+import sopt.org.homepage.project.service.ProjectService;
 import sopt.org.homepage.project.util.ProjectComparator;
 import sopt.org.homepage.project.dto.request.GetProjectsRequestDto;
 import sopt.org.homepage.project.dto.response.ProjectDetailResponseDto;
